+++ settings.gradle.kts
@@ -23,7 +23,6 @@ enableFeaturePreview("IMPROVED_POM_SUPPORT")
 include("distributionsDependencies")
 include("distributions")
 include("baseServices")
-include("baseServicesJava9")
 include("baseServicesGroovy")
 include("logging")
 include("processServices")
