--- subprojects/base-services-java9/base-services-java9.gradle.kts
@@ -1,36 +1,0 @@
-/*
- * Copyright 2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
- * A set of generic services and utilities.
- *
- * Should have a very small set of dependencies, and should be appropriate to embed in an external
- * application (eg as part of the tooling API).
- */
-
-import org.gradle.gradlebuild.unittestandcompile.ModuleType
-import java.util.concurrent.Callable
-
-plugins {
-    `java-library`
-}
-
-java {
-    gradlebuildJava {
-        moduleType = ModuleType.REQUIRES_JAVA_9_COMPILER
-    }
-}
-
